Refactor: decompose ProjectDetail into sibling files
ProjectDetail.tsx was 1354 LOC with six concerns tangled together.
Extract each into its own file so that touching one stops forcing
a scroll through the other five. Intended to be a pure extraction —
no behavior change, no memoization tweaks, no rename of state or
props beyond the doc-mandated component renames.

- src/lib/project-field-formatting.ts: COLOR_TEXT, WORD_OVERRIDES,
  formatFieldKey, formatFieldValue, resolveFieldValue.
- src/components/ProjectRelationshipsTab.tsx: RelationshipsTab ->
  ProjectRelationshipsTab; RelationshipsSidebar ->
  ProjectRelationshipsSidebar (exported); SidebarProjectRow kept
  internal to this file.
- src/components/ProjectEnvironmentsCard.tsx: the environments Card
  block, now parametrized by environments + deploymentStatus.
- src/components/ProjectAttributesSection.tsx: attributeFields useMemo
  + render loop. Renders a fragment of rows so the DOM shape is
  unchanged inside the parent's CardContent.
- src/components/ProjectSettingsTab.tsx: SettingsTab ->
  ProjectSettingsTab. Delete-confirmation modal stays inline in
  this file — separating it is a possible follow-up, not required
  for this PR.

ProjectDetail.tsx is now 599 LOC. PlaceholderTab stays there as a
trivial helper. The noted InlineField memoization / stable onCommit
work is an explicit separate follow-up; untouched here.
